Sanitize hook functions for hooks with :: in them
We should make sure that when we create the _hooks/*.php.twig file
with the code from mw.org hooks, we write the proper function without
the :: or : in there. Capitalize all characters after : or :: and
then remove these completely to create a valid function name that
the code uses.

Fixes #29