Clean up license handling (fixes #4)
Only show the user 3 options: GPL-2.0+, MIT, and Apache-2.0. These
licenses are broad enough to cover most use cases and are the ones
featured on <http://choosealicense.com/>. Including 7 different GPL
versions is probably going to be more confusing than useful :S.

For consistency with MediaWiki, we use 2.0+ instead of 3.0+, and set
that as the default. While MIT may be widespread, it should not be the
default because it does not protect against patent treachery.

Include their license texts as twig templates and create a COPYING file
accordingly.