Redesign; use bootstrap instead of OOUI, add assets to repo
Tool labs server stopped being able to properly update through
npm install with dependencies that are newer than what the server
offers.

To simplify matters, the UI changed from OOUI to bootstrap, and
all assets are now part of the repo so that it can be updated
easily in the production server.

If the repo is forked, make sure to run 'grunt build' if any
of the assets change, and before submitting a pull request.